+++ e2e/cypress/fixtures/images/test-avatar.png
