+++ e2e/cypress/integration/logging.spec.ts
