+++ e2e/cypress/snapshots/base/login.spec.ts/login-form-base.png
