+++ e2e/cypress/snapshots/base/login.spec.ts/login-header-no-avatar-base.png
