+++ e2e/cypress/snapshots/base/login.spec.ts/login-header-with-avatar-and-menu-base.png
