+++ e2e/cypress/snapshots/base/profile.spec.ts/dark-header-base.png
