+++ e2e/cypress/snapshots/base/profile.spec.ts/light-header-base.png
