+++ e2e/cypress/snapshots/base/profile.spec.ts/profile-header-base.png
